+++ lib/src/main/java/com/diffplug/spotless/java/CleanthatStepFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2023 Solven
+ * Copyright 2023 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,11 +15,6 @@
  */
 package com.diffplug.spotless.java;
 
-import com.diffplug.spotless.FormatterFunc;
-import com.diffplug.spotless.FormatterStep;
-import com.diffplug.spotless.JarState;
-import com.diffplug.spotless.Jvm;
-import com.diffplug.spotless.Provisioner;
 import java.io.IOException;
 import java.io.Serializable;
 import java.lang.reflect.Constructor;
@@ -27,9 +22,14 @@ import java.lang.reflect.Method;
 import java.util.List;
 import java.util.Objects;
 
+import com.diffplug.spotless.FormatterFunc;
+import com.diffplug.spotless.FormatterStep;
+import com.diffplug.spotless.JarState;
+import com.diffplug.spotless.Jvm;
+import com.diffplug.spotless.Provisioner;
+
 /**
- * Enables CleanThat as a SpotLess step. This may be moved to Spotless own repo
- * (https://github.com/diffplug/spotless/tree/main/lib)
+ * Enables CleanThat as a SpotLess step.
  *
  * @author Benoit Lacelle
  */
@@ -39,33 +39,45 @@ public final class CleanthatStepFactory {
 	private static final String NAME = "cleanthat";
 	private static final String MAVEN_COORDINATE = "io.github.solven-eu.cleanthat:java:";
 
-	private static final Jvm.Support<String> JVM_SUPPORT = Jvm.<String>support(NAME).add(8, "2.0");
+	private static final Jvm.Support<String> JVM_SUPPORT = Jvm.<String> support(NAME).add(11, "2.0");
 
 	// prevent direct instantiation
-	private CleanthatStepFactory() {
-	}
+	private CleanthatStepFactory() {}
 
-	/** Creates a step which formats everything - code, import order, and unused imports. */
+	/** Creates a step which apply default CleanThat mutators. */
 	public static FormatterStep create(Provisioner provisioner) {
 		return create(defaultVersion(), provisioner);
 	}
 
-	/** Creates a step which formats everything - code, import order, and unused imports. */
+	/** Creates a step which apply default CleanThat mutators. */
 	public static FormatterStep create(String version, Provisioner provisioner) {
-		return create(MAVEN_COORDINATE, version, defaultExcluded(), defaultIncluded(), provisioner);
+		return create(MAVEN_COORDINATE, version, defaultJdkVersion(), defaultExcludedMutators(), defaultMutators(), provisioner);
 	}
 
-	private static List<String> defaultExcluded() {
+	public static String defaultJdkVersion() {
+		// see IJdkVersionConstants.JDK_7
+		// https://maven.apache.org/plugins/maven-compiler-plugin/compile-mojo.html#source
+		// 1.7 is the default for 'maven-compiler-plugin' since 3.9.0
+		return "1.7";
+	}
+
+	public static List<String> defaultExcludedMutators() {
 		return List.of();
 	}
 
-	private static List<String> defaultIncluded() {
+	/**
+	 * By default, we include all available rules
+	 * @return
+	 */
+	public static List<String> defaultMutators() {
+		// see JavaRefactorerProperties.WILDCARD
 		return List.of("*");
 	}
 
-	/** Creates a step which formats everything - groupArtifact, code, import order, and unused imports. */
+	/** Creates a step which apply selected CleanThat mutators. */
 	public static FormatterStep create(String groupArtifact,
 			String version,
+			String sourceJdkVersion,
 			List<String> excluded,
 			List<String> included,
 			Provisioner provisioner) {
@@ -76,8 +88,8 @@ public final class CleanthatStepFactory {
 		Objects.requireNonNull(version, "version");
 		Objects.requireNonNull(provisioner, "provisioner");
 		return FormatterStep.createLazy(NAME,
-				() -> new JavaRulesState(NAME, groupArtifact, version, excluded, included, provisioner),
-				JavaRulesState::createFormat);
+				() -> new JavaRefactorerState(NAME, groupArtifact, version, sourceJdkVersion, excluded, included, provisioner),
+				JavaRefactorerState::createFormat);
 	}
 
 	/** Get default formatter version */
@@ -89,32 +101,35 @@ public final class CleanthatStepFactory {
 		return MAVEN_COORDINATE;
 	}
 
-	static final class JavaRulesState implements Serializable {
+	static final class JavaRefactorerState implements Serializable {
 		private static final long serialVersionUID = 1L;
 
 		final JarState jarState;
 		final String stepName;
 		final String version;
 
+		final String sourceJdkVersion;
 		final List<String> included;
 		final List<String> excluded;
 
-		JavaRulesState(String stepName, String version, Provisioner provisioner) throws IOException {
-			this(stepName, MAVEN_COORDINATE, version, defaultExcluded(), defaultIncluded(), provisioner);
+		JavaRefactorerState(String stepName, String version, Provisioner provisioner) throws IOException {
+			this(stepName, MAVEN_COORDINATE, version, defaultJdkVersion(), defaultExcludedMutators(), defaultMutators(), provisioner);
 		}
 
-		JavaRulesState(String stepName,
+		JavaRefactorerState(String stepName,
 				String groupArtifact,
 				String version,
+				String sourceJdkVersion,
 				List<String> included,
 				List<String> excluded,
 				Provisioner provisioner) throws IOException {
 			JVM_SUPPORT.assertFormatterSupported(version);
-			// ModuleHelper.doOpenInternalPackagesIfRequired();
+			ModuleHelper.doOpenInternalPackagesIfRequired();
 			this.jarState = JarState.from(groupArtifact + ":" + version, provisioner);
 			this.stepName = stepName;
 			this.version = version;
 
+			this.sourceJdkVersion = sourceJdkVersion;
 			this.included = included;
 			this.excluded = excluded;
 		}
@@ -126,11 +141,10 @@ public final class CleanthatStepFactory {
 			Object formatter;
 			Method formatterMethod;
 			try {
-				Class<?> formatterClazz =
-						classLoader.loadClass("com.diffplug.spotless.glue.java.JavaCleanthatRefactoringFunc");
-				Constructor<?> formatterConstructor = formatterClazz.getConstructor(List.class, List.class);
+				Class<?> formatterClazz = classLoader.loadClass("com.diffplug.spotless.glue.java.JavaCleanthatRefactorerFunc");
+				Constructor<?> formatterConstructor = formatterClazz.getConstructor(String.class, List.class, List.class);
 
-				formatter = formatterConstructor.newInstance(included, excluded);
+				formatter = formatterConstructor.newInstance(sourceJdkVersion, included, excluded);
 				formatterMethod = formatterClazz.getMethod("apply", String.class);
 			} catch (ReflectiveOperationException e) {
 				throw new IllegalStateException("Issue executing the formatter", e);